initial organization of src/store
