+++ src/store/README.md
@@ -1,0 +1,6 @@
+## Dependencies
+
+- js-cookie
+- rison
+- mwp-api-state
+- mwp-tracking/util/clickWriter
