+++ src/store/browser/index.js
@@ -1,11 +1,11 @@
 // @flow weak
 import { applyMiddleware, createStore, compose } from 'redux';
-import getClickWriter from '../plugins/tracking/util/clickWriter'; // mwp-tracking/util/clickWriter
-import { getApiMiddleware } from '../api-state'; // mwp-api-state
+import getClickWriter from '../../plugins/tracking/util/clickWriter'; // mwp-tracking/util/clickWriter
+import { getApiMiddleware } from '../../api-state';
 
-import { fetchQueries } from '../util/fetchUtils';
 import catchMiddleware from '../middleware/catch';
 import injectPromise from '../middleware/injectPromise';
+import fetchQueries from './fetchQueries';
 
 declare var document: Object; // ignore 'potentially null' document.body
